[ty] Move the `deferred` submodule inside `infer/builder`
