+++ crates/ty_python_semantic/src/types/infer.rs
@@ -63,7 +63,6 @@ pub(super) use comparisons::UnsupportedComparisonError;
 
 mod builder;
 mod comparisons;
-mod deferred;
 #[cfg(test)]
 mod tests;
