+++ crates/ty_python_semantic/src/types/infer/builder/post_inference/dynamic_class.rs
