+++ crates/ty_python_semantic/src/types/infer/builder/post_inference/static_class.rs
